+++ src/main/kotlin/kr/hhplus/be/server/application/user/UserAccountFacade.kt
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.application.user
+
+class UserAccountFacade {
+}
