+++ src/main/kotlin/kr/hhplus/be/server/domain/user/model/Account.kt
@@ -1,28 +1,28 @@
-package kr.hhplus.be.server.domain.user
+package kr.hhplus.be.server.domain.user.model
 
 import java.time.LocalDateTime
 
-data class Balance private constructor(
+data class Account private constructor(
     val id: Long,
-    val accountId: Long,
+    val userId: Long,
     var amount: Double,
     var createdAt: LocalDateTime,
     var updatedAt: LocalDateTime
 ) {
     companion object {
-        private val MIN_BALANCE: Double = 0.0
-        private val MAX_BALANCE: Double = 100000000.0 // 100,000,000 원
-        private val MAX_TRANSACTION_AMOUNT: Double = 10000000.0 // 10,000,000 원
+        const val MIN_BALANCE: Double = 0.0
+        const val MAX_BALANCE: Double = 100000000.0 // 100,000,000 원
+        const val MAX_TRANSACTION_AMOUNT: Double = 10000000.0 // 10,000,000 원
 
-        fun create(id: Long, accountId: Long, initialAmount: Double = MIN_BALANCE): Balance {
+        fun create(id: Long, userId: Long, initialAmount: Double = MIN_BALANCE): Account {
             require(initialAmount >= MIN_BALANCE && initialAmount <= MAX_BALANCE) {
                 "Initial amount must be between $MIN_BALANCE and $MAX_BALANCE"
             }
-            return Balance(id, accountId, initialAmount, LocalDateTime.now(), LocalDateTime.now())
+            return Account(id, userId, initialAmount, LocalDateTime.now(), LocalDateTime.now())
         }
     }
 
-    fun charge(amount: Double): Balance {
+    fun charge(amount: Double): Account {
         require(amount > 0) { "Charge amount must be positive" }
         require(amount <= MAX_TRANSACTION_AMOUNT) {
             "Charge amount cannot exceed $MAX_TRANSACTION_AMOUNT"
@@ -36,7 +36,7 @@ data class Balance private constructor(
         return this
     }
 
-    fun withdraw(amount: Double): Balance {
+    fun withdraw(amount: Double): Account {
         require(amount > 0) { "Withdrawal amount must be positive" }
         require(amount <= MAX_TRANSACTION_AMOUNT) {
             "Withdrawal amount cannot exceed $MAX_TRANSACTION_AMOUNT"
