+++ src/main/kotlin/kr/hhplus/be/server/domain/user/model/User.kt
@@ -1,8 +1,8 @@
-package kr.hhplus.be.server.domain.user
+package kr.hhplus.be.server.domain.user.model
 
 import java.time.LocalDateTime
 
-data class Account private constructor(
+data class User private constructor(
     val id: Long,
     var name: String,
     var email: String,
@@ -22,32 +22,27 @@ data class Account private constructor(
         private val PASSWORD_REGEX =
             Regex("^(?=.*[A-Za-z])(?=.*\\d)[A-Za-z\\d]{$MIN_PASSWORD_LENGTH,$MAX_PASSWORD_LENGTH}\$")
 
-
         fun create(
             id: Long,
             name: String,
             email: String,
             loginId: String,
             password: String
-        ): Account {
+        ): User {
             require(loginId.length in MIN_LOGIN_ID_LENGTH..MAX_LOGIN_ID_LENGTH) {
                 "Login ID must be between $MIN_LOGIN_ID_LENGTH and $MAX_LOGIN_ID_LENGTH characters"
             }
             require(PASSWORD_REGEX.matches(password)) {
                 "Password must be a combination of letters and numbers and between $MIN_PASSWORD_LENGTH and $MAX_PASSWORD_LENGTH characters"
             }
-            return Account(id, name, email, loginId, password, LocalDateTime.now(), LocalDateTime.now())
+            return User(id, name, email, loginId, password, LocalDateTime.now(), LocalDateTime.now())
         }
     }
 
     fun update(
-        newName: String? = null,
-        newEmail: String? = null,
         newLoginId: String? = null,
         newPassword: String? = null
-    ): Account {
-        newName?.let { this.name = it }
-        newEmail?.let { this.email = it }
+    ): User {
         newLoginId?.let {
             require(it.length in MIN_LOGIN_ID_LENGTH..MAX_LOGIN_ID_LENGTH) {
                 "Login ID must be between $MIN_LOGIN_ID_LENGTH and $MAX_LOGIN_ID_LENGTH characters"
@@ -64,5 +59,4 @@ data class Account private constructor(
         }
         return this
     }
-
 }
