+++ src/main/kotlin/kr/hhplus/be/server/domain/user/service/AccountService.kt
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.domain.user.service
+
+class AccountService {
+}
