+++ src/main/kotlin/kr/hhplus/be/server/interfaces/coupon/dto/request/CouponCreateRequest.kt
