+++ src/main/kotlin/kr/hhplus/be/server/interfaces/coupon/dto/request/CouponUpdateRequest.kt
