+++ src/main/kotlin/kr/hhplus/be/server/interfaces/coupon/dto/response/CouponResponse.kt
