+++ src/main/kotlin/kr/hhplus/be/server/interfaces/product/dto/request/ProductCreateRequest.kt
