+++ src/main/kotlin/kr/hhplus/be/server/interfaces/user/dto/response/UserDetailResponse.kt
